requirements.txt
# Robot control (HTTP API via requests, not opentrons SDK)
requests>=2.31

# Computer vision
opencv-python>=4.8
Pillow>=10.0
super-gradients>=3.7

# Machine learning
numpy>=1.24
scipy>=1.11
scikit-learn>=1.3
botorch>=0.9
gpytorch>=1.11
torch>=2.0

# Data & visualization
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12

# Experiment management
pyyaml>=6.0

# Testing
pytest>=7.0
